# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Todo List API","version":"1.0.0"},"paths":{"\/api\/register":{"post":{"tags":["USER"],"summary":"Register","description":"Register a user","operationId":"App\\Http\\Controllers\\AuthController::register","requestBody":{"description":"JSON Payload","required":true,"content":{"application\/json":{"schema":{"properties":{"name":{"description":"User fullname","type":"string","example":"Api Dev"},"email":{"description":"Unique email address","type":"string","example":"todo@list.com"},"password":{"description":"User password","type":"string","example":""},"password_confirmation":{"description":"Confirm user password","type":"string","example":""}},"type":"object"}}}},"responses":{"200":{"description":"User creation successful.","content":{"application\/json":{"schema":{},"example":{"data":{"id":1,"name":"Api Dev","email":"todo@list.com","created_at":"2019-12-13 16:33:18"}}}}},"422":{"description":"Invalid request parameters.","content":{"application\/json":{"schema":{},"example":{"application\/json":{"error":{"code":422,"message":"The given data was invalid."},"errors":{"email":["The email field is required."]}}}}}}}}},"\/api\/login":{"post":{"tags":["USER"],"summary":"Login","description":"Login user","operationId":"App\\Http\\Controllers\\AuthController::login","requestBody":{"description":"JSON Payload","required":true,"content":{"application\/json":{"schema":{"properties":{"email":{"description":"User email address","type":"string","example":"todo@list.com"},"password":{"description":"User password","type":"string","example":""}},"type":"object"}}}},"responses":{"200":{"description":"User creation successful.","content":{"application\/json":{"schema":{},"example":{"data":{"success":true,"token":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9hcGkudG9kby50ZXN0XC9hcGlcL2xvZ2luIiwiaWF0IjoxNTc2MjI1NjM1LCJleHAiOjE2MDc3NjE2MzUsIm5iZiI6MTU3NjIyNTYzNSwianRpIjoiWTRHSlYxSEpYa21UaTdYbCIsInN1YiI6MTksInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.-LUG9Yqtat3ZR0EJ0PIs6ExObFdLBfGi3pQJ_lckw3Q","user":{"id":19,"name":"Api Dev","email":"ace@list.com","created_at":"2019-12-13 08:19:59","updated_at":"2019-12-13 08:19:59","deleted_at":null}}}}}},"422":{"description":"Invalid request parameters.","content":{"application\/json":{"schema":{},"example":{"application\/json":{"error":{"code":422,"message":"The given data was invalid."},"errors":{"email":["The email field is required."]}}}}}}}}},"\/api\/todo":{"get":{"tags":["TODO"],"summary":"Get all todo of a user","description":"Fetch all available todo list.","operationId":"App\\Http\\Controllers\\TodoController::index","responses":{"200":{"description":"","content":{"application\/json":{"schema":{},"example":{"data":[[{"id":1,"title":"test","description":"asdads","is_done":"asdads","created_at":"2019-12-13 10:46:39"},{"id":2,"title":"qwe","description":"eewqeqe","is_done":"eewqeqe","created_at":"2019-12-13 10:46:39"}]]}}}},"401":{"description":"Unauthenticated.","content":{"application\/json":{"schema":{},"example":{"error":{"code":401,"message":"Unauthenticated."}}}}},"403":{"description":"Unauthorized.","content":{"application\/json":{"schema":{},"example":{"error":{"code":403,"message":"Unauthorized."}}}}},"404":{"description":"User not found.","content":{"application\/json":{"schema":{},"example":{"error":{"code":404,"message":"Record not found."}}}}}},"security":[{"bearer":[]}]},"post":{"tags":["TODO"],"summary":"Create todo","description":"Create todo","operationId":"App\\Http\\Controllers\\TodoController::store","requestBody":{"description":"JSON Payload","required":true,"content":{"application\/json":{"schema":{"properties":{"title":{"description":"Title of the todo","type":"string","example":""},"description":{"description":"Description of the todo","type":"string","example":""},"is_done":{"description":"","type":"bolean","example":false}},"type":"object"}}}},"responses":{"200":{"description":"Todo creation successful.","content":{"application\/json":{"schema":{},"example":{"data":[]}}}},"422":{"description":"Invalid request parameters.","content":{"application\/json":{"schema":{},"example":{"application\/json":{"error":{"code":422,"message":"The given data was invalid."},"errors":{"email":["The title field is required."]}}}}}}},"security":[{"bearer":[]}]}},"\/api\/todo\/{todo_id}":{"get":{"tags":["TODO"],"summary":"Get todo by id","description":"Get a todo by id","operationId":"App\\Http\\Controllers\\TodoController::show","parameters":[{"name":"todo_id","in":"path","description":"Todo id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Fetch successful.","content":{"application\/json":{"schema":{},"example":{"data":{"id":11,"title":"updated","description":"updatetete","is_done":true,"created_at":"2019-12-16 05:00:07"}}}}}},"security":[{"bearer":[]}]},"delete":{"tags":["TODO"],"summary":"Delete todo","description":"Delete a todo","operationId":"App\\Http\\Controllers\\TodoController::destroy","parameters":[{"name":"todo_id","in":"path","description":"Todo id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Delete successful.","content":{"application\/json":{"schema":{},"example":{"data":[]}}}}},"security":[{"bearer":[]}]},"patch":{"tags":["TODO"],"summary":"Update todo","description":"Update a todo","operationId":"App\\Http\\Controllers\\TodoController::update","parameters":[{"name":"todo_id","in":"path","description":"Todo id","required":true,"schema":{"type":"integer"}}],"requestBody":{"description":"JSON Payload","required":true,"content":{"application\/json":{"schema":{"properties":{"title":{"description":"Title of the todo","type":"string","example":""},"description":{"description":"Description of the todo","type":"string","example":""},"is_done":{"description":"","type":"bolean","example":false}},"type":"object"}}}},"responses":{"200":{"description":"Update todo successful.","content":{"application\/json":{"schema":{},"example":{"data":{"id":11,"title":"updated","description":"updatetete","is_done":true,"created_at":"2019-12-16 05:00:07"}}}}},"422":{"description":"Invalid request parameters.","content":{"application\/json":{"schema":{},"example":{"application\/json":{"error":{"code":422,"message":"The given data was invalid."},"errors":{"email":["The title field is required."]}}}}}}},"security":[{"bearer":[]}]}},"\/api\/todo\/{todo_id}\/status":{"patch":{"tags":["TODO"],"summary":"Update todo status","description":"Update a todo status","operationId":"App\\Http\\Controllers\\TodoController::updateStatus","parameters":[{"name":"todo_id","in":"path","description":"Todo id","required":true,"schema":{"type":"integer"}}],"requestBody":{"description":"JSON Payload","required":true,"content":{"application\/json":{"schema":{"properties":{"status":{"description":"true or false","type":"boolean","example":"true"}},"type":"object"}}}},"responses":{"200":{"description":"Update todo status successful.","content":{"application\/json":{"schema":{},"example":{"data":{"id":11,"title":"updated","description":"updatetete","is_done":true,"created_at":"2019-12-16 05:00:07"}}}}}},"security":[{"bearer":[]}]}}},"components":{"securitySchemes":{"bearer":{"type":"apiKey","description":"Bearer type authorization","name":"Authorization","in":"header"}}}}